Backend/requirements.txt
Django<=4.2.6
djangorestframework-jsonapi==6.0.0
django-cors-headers~=4.3.0
djangorestframework~=3.14.0
django-environ~=0.11.2
beautifulsoup4==4.12.2
django-cors-headers==4.3.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
django-db-connection-pool==1.2.4
uvicorn==0.23.2
drf-spectacular>=0.15.1,<0.16
django-extensions==3.2.1
requests==2.31.0
whitenoise==6.6.0